+++ tests/data/pyink/module_docstring_2.py
@@ -1,0 +1,4 @@
+"""This is the first line.
+
+This is another paragraph line that perfectly fits in the line length---- limit.
+"""